+++ TabloidMVC/Repositories/IUserProfileRepository.cs
@@ -10,11 +10,11 @@ namespace TabloidMVC.Repositories
         UserProfile GetByEmail(string email);
         UserProfile GetById(int id);
         int GetAdminCount();
-        int GetRequesterId(int userId, byte requestType);
-        byte RequestsByUserId(int id, byte requestType);
+        int GetRequesterId(int userId, int requestType);
+        int RequestsByUserId(int id, int requestType);
         void RequestChange(AdminRequest request);
-        void RetireRequest(int userId, byte requestType);
-        void CancelRequest(int requesterId, int userId, byte requestType);
+        void RetireRequest(int userId, int requestType);
+        void CancelRequest(int requesterId, int userId, int requestType);
         void Add(UserProfile userProfile);
         void Deactivate (UserProfile userProfile);
         void Activate (UserProfile userProfile);
